+++ src/Order/Service/OrderService.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
 namespace FreshAdvance\Invoice\Order\Service;
 
 use FreshAdvance\Invoice\DataType\InvoiceDataInterface;
-use FreshAdvance\Invoice\Repository\OrderRepositoryInterface;
+use FreshAdvance\Invoice\Order\Repository\OrderRepositoryInterface;
 
 class OrderService implements OrderServiceInterface
 {