+++ tests/Unit/Order/Service/OrdereServiceTest.php
@@ -10,9 +10,9 @@ declare(strict_types=1);
 namespace FreshAdvance\Invoice\Tests\Unit\Order\Service;
 
 use FreshAdvance\Invoice\DataType\InvoiceDataInterface;
+use FreshAdvance\Invoice\Order\Repository\OrderRepositoryInterface;
 use FreshAdvance\Invoice\Order\Service\OrderService;
 use FreshAdvance\Invoice\Order\Service\OrderServiceInterface;
-use FreshAdvance\Invoice\Repository\OrderRepositoryInterface;
 use OxidEsales\Eshop\Application\Model\Order;
 use PHPUnit\Framework\TestCase;